+++ core/modules/modReferenceLetters.class.php
@@ -62,6 +62,7 @@ class modReferenceLetters extends DolibarrModules
 		$this->description = 'DocEdit (a.k.a. ReferenceLetters) allows you to create PDF templates using a wysiwyg editor';
 		// Possible values for version are: 'development', 'experimental' or version
 
+
 		$this->version = '2.21.0';
 		// Url to the file with your last numberversion of this module
 		require_once __DIR__ . '/../../class/techatm.class.php';