+++ ChangeLog.md
@@ -17,6 +17,7 @@ All notable changes to this project will be documented in this file.
 *16/04/2024* - 2.20.0  
 
 ## Release 2.19 - 12/02/2024
+- FIX : DA024939 - Handle break page with subtotal - *17/05/2024* - 2.19.7
 - FIX : DA024875 some keys were missing  - *02/05/2024* - 2.19.6
 - FIX : DA024820 trainee extrafields wasn't working properly  - *02/05/2024* - 2.19.5
 - FIX : DA024774 missing phone_perso & phone_mobile key  - *23/04/2024* - 2.19.3